Move MediaFramework files to framework
Modify cxxtest/Make.defs and framework/Makefile
