--- framework/include/media/InputDataSource.hpp
@@ -1,11 +1,0 @@
-#pragma once
-
-#include <media/DataSource.hpp>
-
-class InputDataSource : public DataSource
-{
-public:
-	InputDataSource();
-	~InputDataSource();
-};
-
