--- framework/include/media/OutputDataSource.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "DataSource.h"
-
-class OutputDataSource : public DataSource
-{
-public:
-	OutputDataSource();
-	~OutputDataSource();
-};
