--- framework/include/media/media_types.h
@@ -1,37 +1,0 @@
-/****************************************************************************
- *
- * Copyright 2017 Samsung Electronics All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific
- * language governing permissions and limitations under the License.
- *
- ****************************************************************************/
-
-#ifndef __MEDIA_TYPES_H__
-#define __MEDIA_TYPES_H__
-
-/***
- * Errors
- */
-#define MEDIA_OK						0
-#define MEDIA_ERROR						1
-#define MEDIA_ERROR_NOT_READY			2
-#define MEDIA_ERROR_INVALID_PARAMETER	3
-
-enum media_format_e {
-	MEDIA_FORMAT_PCM,
-	MEDIA_FORMAT_WAV
-};
-
-typedef enum media_format_e media_format_t;
-
-#endif
