+++ framework/src/media/Make.defs
@@ -18,16 +18,12 @@
 
 ifeq ($(CONFIG_MEDIA), y)
 ifeq ($(CONFIG_MEDIA_PLAYER), y)
-#CSRCS += media_player.c
 CXXSRCS += MediaPlayer.cxx
 endif
 
 ifeq ($(CONFIG_MEDIA_RECORDER), y)
-#CSRCS += media_recorder.c
 CXXSRCS += MediaRecorder.cxx
 endif
-
-#CSRCS += media_utils.c
 
 DEPPATH += --dep-path src/media
 VPATH += :src/media
