--- framework/src/media/OutputDataSource.cxx
@@ -1,11 +1,0 @@
-#include "OutputDataSource.h"
-
-
-OutputDataSource::OutputDataSource()
-{
-}
-
-
-OutputDataSource::~OutputDataSource()
-{
-}
